[SID:2505d27d] fix(presence): presence FAB toggle + working-count badge (선생님 가시성)
선생님 dev 피드백: <2xl 토글(우측 가장자리 탭)이 "정신병 수준" 가시성(안 보임)·접근 죽으면 패널 가치(한눈에) 죽음. 선생님 안=우하단 FAB.

- 우측 edge 탭 → **우하단 FAB**(fixed bottom-6 right-6[모바일 bottom-20 GNB 회피]·size-14·rounded-full·bg-brand·shadow-lg·Users 22px·동적 aria/tooltip). 전 width 명확 가시.
- **working-count 배지**(FAB 우상단·"N") — 패널 닫힘 상태서도 "N 작업 중" at-a-glance(선생님 의도 강화). 의미=지금 채팅 답 생성 중 N명(현 working 신호 한정·PO 인지).
- **폴 상향**: 신규 useTeamPresence hook(단일 폴 ~3s·document.hidden pause) → ScrollShell에서 FAB 배지 + 패널 둘 다 공급(중복0). 패널은 items prop 수신(내부 폴 제거). doc §6 "닫힘 시 중단"은 배지 요구로 supersede.
- i18n presence.fabLabelWorking(en/ko)·매직hex0(brand/brand-foreground).

type-check·lint(변경0)·build 그린.
